add automatic installation of missing plugins also for esm_runscripts
